+++ src/Laraplans/Contracts/PlanSubscriptionInterface.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Gerardojbaez\LaraPlans\Contracts;
+namespace Gerardojbaez\Laraplans\Contracts;
 
 interface PlanSubscriptionInterface
 {
-    public function user();
+    public function subscribable();
     public function plan();
     public function usage();
     public function getStatusAttribute();
-    public function active();
+    public function isActive();
     public function onTrial();
-    public function canceled();
-    public function ended();
+    public function isCanceled();
+    public function isEnded();
     public function renew();
     public function cancel($immediately);
     public function changePlan($plan);