+++ src/database/factories/PlanSubscriptionFactory.php
@@ -1,12 +1,13 @@
 <?php
 
-use Gerardojbaez\LaraPlans\Models\Plan;
-use Gerardojbaez\LaraPlans\Tests\Models\User;
-use Gerardojbaez\LaraPlans\Models\PlanSubscription;
+use Gerardojbaez\Laraplans\Models\Plan;
+use Gerardojbaez\Laraplans\Tests\Models\User;
+use Gerardojbaez\Laraplans\Models\PlanSubscription;
 
 $factory->define(PlanSubscription::class, function (Faker\Generator $faker) {
     return [
-        'user_id' => factory(User::class)->create()->id,
+        'subscribable_id' => factory(User::class)->create()->id,
+        'subscribable_type' => User::class,
         'plan_id' => factory(Plan::class)->create()->id,
         'name' => $faker->word
     ];