+++ tests/integration/Models/PlanSubscriptionUsageTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Gerarodjbaez\LaraPlans\Tests\Integration\Models;
+namespace Gerarodjbaez\Laraplans\Tests\Integration\Models;
 
 use Config;
 use Carbon\Carbon;
-use Gerardojbaez\LaraPlans\Models\Plan;
-use Gerardojbaez\LaraPlans\Tests\TestCase;
-use Gerardojbaez\LaraPlans\Tests\Models\User;
-use Gerardojbaez\LaraPlans\Models\PlanFeature;
+use Gerardojbaez\Laraplans\Models\Plan;
+use Gerardojbaez\Laraplans\Tests\TestCase;
+use Gerardojbaez\Laraplans\Tests\Models\User;
+use Gerardojbaez\Laraplans\Models\PlanFeature;
 
 class PlanSubscriptionUsageTest extends TestCase
 {
@@ -21,8 +21,8 @@ class PlanSubscriptionUsageTest extends TestCase
     {
         Config::set('laraplans.features', [
             'listings_per_month' => [
-                'reseteable_interval' => 'month',
-                'reseteable_count' => 1
+                'resettable_interval' => 'month',
+                'resettable_count' => 1
             ]
         ]);