+++ CHANGELOG.md
@@ -3,6 +3,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+*This release breaks backward compatibility.*
+
+### Added
+- `SubscriptionRenewed` event.
+- `SubscriptionCanceled` event.
+- `SubscriptionPlanChanged` event.
+
+### Changed
+- Namespace changed from `Gerardojbaez\LaraPlans` to `Gerardojbaez\Laraplans`
+
+### Fixed
+- Fix #18
+- Fix #17
+- Fix #11
+
 ## [1.0.0] - 2016-03-05
 ### Added
 - This change log file